Add command line options
Finally give the program a real UI. The usage is
based on that of the Windows jaggd.exe program,
with one tweak: Hex numbers can be specified as
0x<num> in addition to $<num> for convenience.

Including a reset option on the command line still
causes the program to exit after reseting. That
will be cleaned up in a subsequent change.